// packages/api-client/src/schema.d.ts
export interface components {
    schemas: {
// ...
        Job: {
            /** Format: uuid */
            id: string;
            status: components["schemas"]["JobStatusEnum"];
            result: unknown;
            error: string | null;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}